Can now Add, Update & Delete MULTIPLE Contacts to/from Users
- additionally `api-routes.js` refactored for cleaner code

+++ backend/controllers/contactController.js
@@ -26,33 +26,51 @@ exports.index = function (req, res) {
 
 // POST - create new contact
 exports.new = function (req, res) {
-    let contact = new Contact();
-    contact.firstName = req.body.firstName ? req.body.firstName : contact.firstName;
-    contact.avatarImage = req.body.avatarImage;
-    contact.email = req.body.email;
-    contact.relation = req.body.relation;
-
     User.findById(req.params.user_id, function (err, user) {
         if (err) {
             res.json(err);
         }
-        if (user != null) {
-            user.Contacts.push(contact);
-            user.save(function () {
-                contact.save(function (err) {
-                    if (err) {
-                        res.json(err);
-                    }
-                    res.json({
-                        message: 'Contact created and added to ' + user.firstName + '!',
-                        data: contact
+        // FIRST check if User exists
+        if (user) {
+            // SECOND, check if Contact exists in Contact model/schema
+            let contact;
+            Contact.findById(req.body._id, function (err, cont) {
+                if (!cont) {
+                    contact = new Contact();
+                    contact.firstName = req.body.firstName ? req.body.firstName : contact.firstName;
+                    contact.avatarImage = req.body.avatarImage;
+                    contact.email = req.body.email;
+                    contact.relation = req.body.relation;
+
+                    contact.save(function (err) {
+                        if (err) {
+                            res.json(err);
+                        }
                     });
-                });
+                } else {
+                    contact = cont;
+                }
+
+                // THIRDLY, if User already contains Contact with same id then don't add a new one
+                if (!user.Contacts.find(cont => cont._id === contact._id)) {
+                    user.Contacts.push(contact);
+                    user.save(function () {
+                        if (err) {
+                            res.json(err);
+                        }
+                        res.json({
+                            message: 'Contact added to ' + user.firstName + '!',
+                            data: contact
+                        });
+                    });
+                } else {
+                    res.json({message: user.firstName + " already has " + contact.firstName + " in their Contacts!"});
+                }
             });
         }
         // to check when user is null
         else {
-            res.json({message: "ERROR"});
+            res.json({message: "ERROR: User does not exist!"});
         }
     });
 };
@@ -79,24 +97,24 @@ exports.update = function (req, res) {
 
         // save the contact in DB and check for errors
         contact.save(function () {
-            // finds ONE user that has a Contact with `contact_id` in it
-            // (TO-DO: might need to change `findOne` later when multiple users have same Contact)
-            User.findOne({"Contacts._id": contact._id}, function (err, user) {
-                // need index so that we change the correct User in the array
-                let index = user.Contacts.findIndex(contact2 => contact2._id === contact._id);
-
+            User.find({"Contacts._id": contact._id}, function (err, users) {
                 if (err) {
                     res.json(err);
                 } else {
-                    user.Contacts[index] = contact;
+                    users.forEach(user => {
+                        // need index so that we change the correct User in the array
+                        let index = user.Contacts.findIndex(contact2 => contact2._id === contact._id);
 
-                    user.save(function (err) {
-                        if (err)
-                            res.json(err);
-                        res.json({
-                            message: 'Contact Info updated',
-                            data: user
+                        user.Contacts[index] = contact;
+
+                        user.save(function (err) {
+                            if (err)
+                                res.json(err);
                         });
+                    });
+                    res.json({
+                        message: 'Contact Info updated',
+                        data: users
                     });
                 }
             });
@@ -108,28 +126,45 @@ exports.update = function (req, res) {
 exports.delete = function (req, res) {
     Contact.findById(req.params.contact_id, function (err, contact) {
         if (contact) {
-            User.findOne({"Contacts._id": contact._id}, function (err, user) {
-                // First, remove Contact from User's "Contacts" array
-                user.Contacts.remove(contact);
-                user.save(function () {
-                    // Secondly, remove Contact from actual Contact schema
-                    Contact.deleteOne({
-                        _id: req.params.contact_id
-                    }, function (err) {
-                        if (err)
-                            res.send(err);
-                        res.json({
-                            status: "success",
-                            message: 'Contact deleted',
-                            data: user
+            User.findById(req.params.user_id, function (err, user) {
+                if (user && (user.Contacts.find(cont => cont._id === contact._id))) {
+                    // First, remove Contact from User's "Contacts" array
+                    user.Contacts.remove(contact);
+
+                    user.save(function (err) {
+                        if (err) {
+                            res.json(err);
+                        }
+                        // Delete contact from Contacts when no User is using that Contact
+                        User.find({"Contacts._id": contact._id}, function (err, users) {
+                            if (!users.length) {
+                                // Secondly, remove Contact from actual Contact schema
+                                Contact.deleteOne({
+                                    _id: req.params.contact_id
+                                }, function (err) {
+                                    if (err)
+                                        res.send(err);
+                                    res.json({
+                                        status: "success",
+                                        message: 'Contact deleted from ' + user.firstName + ' and DB.',
+                                        data: user
+                                    });
+                                });
+                            } else {
+                                res.json('Contact removed from ' + user.firstName + ".");
+                            }
                         });
                     });
-                });
+                } else {
+                    res.json("ERROR: User does not exist or does not have " + contact.firstName + " as a Contact.");
+                }
             });
+
+
         } else {
             res.json({
                 message: "ERROR: Contact does not exist! Perhaps you entered a User ID?"
-            })
+            });
         }
     });
 };
